% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAcons.R
\docType{data}
\name{AACons}
\alias{AACons}
\title{Amino acids at highly conserved position in HLA sequences.}
\format{object of class RData. List of characters.}
\usage{
data(AACons)
}
\description{
List of Amino acids of 10 positions highly conserved in large dataset of alpha1 and alpha2 domain sequences. Theses positions and amino acids had conservation
=> 99.8 % based on 9450 HLA-alleles sequences from IMGT/HLA (Robinson et al. 2010).
}
\keyword{datasets}
