% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_lm.R
\docType{data}
\name{model_lm}
\alias{model_lm}
\title{Linear regression model}
\format{object of class RData}
\usage{
data(model_lm)
}
\description{
Linear regression model for each most commun HLA-gene (A,B and C) trained exclusively on Mass spectrometry dataset from multi-allelics samples. (Steffe et al. 2019)
}
\keyword{datasets}
