% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_pos.R
\docType{data}
\name{model_pos}
\alias{model_pos}
\title{HLA sequence positions selection}
\format{object of class RData}
\usage{
data(model_pos)
}
\description{
HLA sequences positions related to ligand length distribution (correlation of less than -0.3 or superior to 0.3) selected based on ligand length distribution of multi-allelic samples (Steffe et al. 2019)
}
\keyword{datasets}
