% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_dis.R
\name{norm_dis}
\alias{norm_dis}
\title{normalisation ligand length distribution}
\usage{
norm_dis(x)
}
\arguments{
\item{x}{ligand length distribution to normalise}
}
\value{
Vector with ligand length distribution normalised.
}
\description{
Normalisation of ligand length distribution when prediction of one of ligang length is negative.
}
