% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosCons.R
\docType{data}
\name{PosCons}
\alias{PosCons}
\title{Position to define domaine a1 and a2 of HLA molecule}
\format{object of class RData. List of interger.}
\usage{
data(PosCons)
}
\description{
Vector of 10 positions highly conserved in all HLA sequences. The 10 positions had conservation score of => 99.8 % based on 9450 HLA-allele from IMGT/HLA (Robinson et al. 2010)
}
\keyword{datasets}
