% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PWMgenes.R
\docType{data}
\name{PosWM}
\alias{PosWM}
\title{Position Weigth Matrix per HLA-gene}
\format{object of class RData}
\usage{
data(PWMgenes)
}
\description{
Position weight matrix calculated for each gene based on HLA sequences in IMGT/HLA (Robinson et al. 2010).
}
\keyword{datasets}
