% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_LD.R
\name{Prediction_LD}
\alias{Prediction_LD}
\title{Prediction of HLA class I ligand length distribution}
\usage{
Prediction_LD(empty.seq, sequences, allele, gene)
}
\arguments{
\item{empty.seq}{HLA without a1 and a2 domains found in sequence.}

\item{sequences}{HLA sequences (vector or dataframe)}

\item{allele}{Allele of HLA (e.g.'HLA-A01:01')}

\item{gene}{Gene of HLA}
}
\value{
dataframe with gene/allele and the corresponding ligand length distribution predicted with linear regression model (Save in csv format).
}
\description{
Prediction_LD allows the prediction of HLA-gene if it is unknown based on position weight matrix.
}
