% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScorePWM.R
\name{ScorePWM}
\alias{ScorePWM}
\title{Score sequence similarity per HLA-gene}
\usage{
ScorePWM(sequences, PWMgene)
}
\arguments{
\item{sequences}{HLA-I amino acids sequences}

\item{PWMgene}{Position Weight Matrix for each HLA-gene}
}
\value{
Score of sequences similarity with HLA-gene
}
\description{
Sum of amino acids frequencies at specific position in the gene-specific PWMs corresponding to sequences.
}
