% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_definition.R
\name{sequence_definition}
\alias{sequence_definition}
\title{Delimitation of HLA sequences for prediction}
\usage{
sequence_definition(seq)
}
\arguments{
\item{seq}{protein sequence of HLAs}
}
\description{
Define a1 and a2 domains in sequence using 10 highly conserved positions.
}
