% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_format.R
\name{sequence_format}
\alias{sequence_format}
\title{Sequence format}
\usage{
sequence_format(sequences)
}
\arguments{
\item{sequences}{list of amino acids}
}
\value{
sequences corrected from automatic correction for 'T' and 'F'
}
\description{
Correct the automatic conversion of "T" to TRUE or "F" to FALSE.
}
