% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sequence_vector.R
\name{Sequence_vector}
\alias{Sequence_vector}
\title{Sequence in 20 dimension vectors}
\usage{
Sequence_vector(Sequence)
}
\arguments{
\item{Sequence}{vector or dataframe of HLA sequences.}
}
\value{
dataframe where each line represent sequences coded as 20 dimensions vector per position and ligand length distribution for each allele if pHF not null
}
\description{
Create a one-hot encoding scheme of HLA sequences (per 20 amino-acids).
}
