% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.test.R
\docType{data}
\name{sequences.test}
\alias{sequences.test}
\title{HLA-I sequences for packages test}
\format{object of class RData. List of interger.}
\usage{
data(sequences.test)
}
\description{
6 sequences from IMGT/HLA (Robinson et al. 2015).
}
\keyword{datasets}
