% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction_LD.R
\name{Prediction_LD}
\alias{Prediction_LD}
\title{Prediction of HLA class I ligand length distribution}
\usage{
Prediction_LD(empty.seq, sequences, pred_sparse, pos_cor, model_pos,
  model_lm, allele, gene)
}
\arguments{
\item{empty.seq}{HLA with  no a1 and a2 sequences found.}

\item{sequences}{HLA sequences (vector or dataframe)}

\item{pred_sparse}{Sparse matrix of HLA sequences}

\item{pos_cor}{selection of position for the prediction}

\item{model_pos}{position for linear regression model for each of HLA-gene (HLA-A,B & C)}

\item{model_lm}{linear regression trained on multi-allelic samples per HLA-gene (HLA-A,B & C)}

\item{allele}{allele of HLA (e.g.'HLA-A01:01')}

\item{gene}{gene of HLA}
}
\value{
dataframe with gene/allele and the corresponding ligand length distribution predicted with linear regression model (Save in csv format).
}
\description{
Prediction_LD allows the prediction of HLA-gene if it is unknown based on position weight matrix.
}
