% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_definition.R
\name{sequence_definition}
\alias{sequence_definition}
\title{Definition of sequence HLA sequences}
\usage{
sequence_definition(seq)
}
\arguments{
\item{seq}{protein sequence of HLAs}
}
\description{
Define sequences a1 abd a2 as described in the paper Steffe et al. 2019 using 10 highly conserved positions.
}
